+++ ncgen/util.c
@@ -28,13 +28,13 @@ chartohex(char c)
     switch (c) {
         case '0': case '1': case '2': case '3': case '4':
         case '5': case '6': case '7': case '8': case '9':
-            return (c - '0');
+            return (unsigned int)(c - '0');
         case 'A': case 'B': case 'C':
         case 'D': case 'E': case 'F':
-            return (c - 'A') + 0x0a;
+            return (unsigned int)(c - 'A') + 0x0a;
         case 'a': case 'b': case 'c':
         case 'd': case 'e': case 'f':
-            return (c - 'a') + 0x0a;
+            return (unsigned int)(c - 'a') + 0x0a;
     }
     return 0;
 }
